+++ contracts/BonsaiStarterLowLevel.sol
@@ -24,7 +24,7 @@ import {BonsaiLowLevelCallbackReceiver} from "bonsai/BonsaiLowLevelCallbackRecei
 //       or difficult to implement function to a RISC Zero guest running on Bonsai.
 contract BonsaiStarterLowLevel is BonsaiLowLevelCallbackReceiver {
     // Cache of the results calculated by our guest program in Bonsai.
-    mapping(uint256 => uint256) public fibonacciCache;
+    mapping(uint32 => uint256) public fibonacciCache;
 
     // The image id of the only binary we accept callbacks from
     bytes32 public immutable fibImageId;
@@ -39,7 +39,7 @@ contract BonsaiStarterLowLevel is BonsaiLowLevelCallbackReceiver {
     /// @notice Returns nth number in the Fibonacci sequence.
     /// @dev The sequence is defined as 1, 1, 2, 3, 5 ... with fibonacci(0) == 1.
     ///      Only precomputed results can be returned. Call calculate_fibonacci(n) to precompute.
-    function fibonacci(uint256 n) external view returns (uint256) {
+    function fibonacci(uint32 n) external view returns (uint256) {
         uint256 result = fibonacciCache[n];
         require(result != 0, "value not available in cache");
         return result;
@@ -48,7 +48,7 @@ contract BonsaiStarterLowLevel is BonsaiLowLevelCallbackReceiver {
     /// @notice Callback function logic for processing verified journals from Bonsai.
     function bonsaiLowLevelCallback(bytes calldata journal, bytes32 imageId) internal override returns (bytes memory) {
         require(imageId == fibImageId);
-        (uint256 n, uint256 result) = abi.decode(journal, (uint256, uint256));
+        (uint32 n, uint256 result) = abi.decode(journal, (uint32, uint256));
         emit CalculateFibonacciCallback(n, result);
         fibonacciCache[n] = result;
         return new bytes(0);
@@ -58,7 +58,7 @@ contract BonsaiStarterLowLevel is BonsaiLowLevelCallbackReceiver {
     /// @dev This function sends the request to Bonsai through the on-chain relay.
     ///      The request will trigger Bonsai to run the specified RISC Zero guest program with
     ///      the given input and asynchronously return the verified results via the callback below.
-    function calculateFibonacci(uint256 n) external {
+    function calculateFibonacci(uint32 n) external {
         bonsaiRelay.requestCallback(
             fibImageId, abi.encode(n), address(this), this.bonsaiLowLevelCallbackReceiver.selector, 30000
         );
